+++ lib/commercetools-api/src/Models/OrderEdit/StagedOrderModel.php
@@ -260,6 +260,12 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
 
     /**
      *
+     * @var ?string
+     */
+    protected $purchaseOrderNumber;
+
+    /**
+     *
      * @var ?DiscountCodeInfoCollection
      */
     protected $discountCodes;
@@ -374,6 +380,7 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
         ?ShippingInfo $shippingInfo = null,
         ?SyncInfoCollection $syncInfo = null,
         ?ReturnInfoCollection $returnInfo = null,
+        ?string $purchaseOrderNumber = null,
         ?DiscountCodeInfoCollection $discountCodes = null,
         ?int $lastMessageSequenceNumber = null,
         ?CartReference $cart = null,
@@ -421,6 +428,7 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
         $this->shippingInfo = $shippingInfo;
         $this->syncInfo = $syncInfo;
         $this->returnInfo = $returnInfo;
+        $this->purchaseOrderNumber = $purchaseOrderNumber;
         $this->discountCodes = $discountCodes;
         $this->lastMessageSequenceNumber = $lastMessageSequenceNumber;
         $this->cart = $cart;
@@ -1101,6 +1109,27 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
     }
 
     /**
+     * <p>The Purchase Order Number is typically set by the <a href="/quotes-overview#buyer">Buyer</a> on a <a href="ctp:api:type:QuoteRequest">QuoteRequest</a> to
+     * track the purchase order during the <a href="/../api/quotes-overview#intended-workflow">quote and order flow</a>.</p>
+     *
+     *
+     * @return null|string
+     */
+    public function getPurchaseOrderNumber()
+    {
+        if (is_null($this->purchaseOrderNumber)) {
+            /** @psalm-var ?string $data */
+            $data = $this->raw(self::FIELD_PURCHASE_ORDER_NUMBER);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->purchaseOrderNumber = (string) $data;
+        }
+
+        return $this->purchaseOrderNumber;
+    }
+
+    /**
      *
      * @return null|DiscountCodeInfoCollection
      */
@@ -1617,6 +1646,14 @@ final class StagedOrderModel extends JsonObjectModel implements StagedOrder
     public function setReturnInfo(?ReturnInfoCollection $returnInfo): void
     {
         $this->returnInfo = $returnInfo;
+    }
+
+    /**
+     * @param ?string $purchaseOrderNumber
+     */
+    public function setPurchaseOrderNumber(?string $purchaseOrderNumber): void
+    {
+        $this->purchaseOrderNumber = $purchaseOrderNumber;
     }
 
     /**