+++ lib/commercetools-api/src/Models/Quote/Quote.php
@@ -65,6 +65,7 @@ interface Quote extends BaseResource
     public const FIELD_CUSTOM = 'custom';
     public const FIELD_QUOTE_STATE = 'quoteState';
     public const FIELD_STATE = 'state';
+    public const FIELD_PURCHASE_ORDER_NUMBER = 'purchaseOrderNumber';
     public const FIELD_BUSINESS_UNIT = 'businessUnit';
 
     /**
@@ -349,6 +350,15 @@ interface Quote extends BaseResource
     public function getState();
 
     /**
+     * <p>The Purchase Order Number is typically set by the <a href="/quotes-overview#buyer">Buyer</a> on a <a href="ctp:api:type:QuoteRequest">QuoteRequest</a> to
+     * track the purchase order during the <a href="/../api/quotes-overview#intended-workflow">quote and order flow</a>.</p>
+     *
+
+     * @return null|string
+     */
+    public function getPurchaseOrderNumber();
+
+    /**
      * <p>The <a href="ctp:api:type:BusinessUnit">BusinessUnit</a> for the Quote.</p>
      *
 
@@ -525,6 +535,11 @@ interface Quote extends BaseResource
      * @param ?StateReference $state
      */
     public function setState(?StateReference $state): void;
+
+    /**
+     * @param ?string $purchaseOrderNumber
+     */
+    public function setPurchaseOrderNumber(?string $purchaseOrderNumber): void;
 
     /**
      * @param ?BusinessUnitKeyReference $businessUnit